startxfce4: mark .desktop links in Desktop as safe
The XFCE version in EL9 introduced that .desktop files outside of `/usr/share` etc.
(XDG_DATA_DIRS) need to be acknowledged by the user as safe before running.
Their SHA256 needs to be written to a GVFS metadata value.

(This could also go to the entrypoint.sh but then needs to be run as the container user.)